Re-download file missing from cache
The problem was that the etag value was specified and the server can return
http.StatusNotModified in which case the missing file is not downloaded. But
even if the file was downloaded, e.g. because its etag value had changed, an
empty path would still have been returned.

Also the downloading messages need to be written to the correct destination
for consistency with the rest of the code.